build: generate vmlinux at compile time
Previously SeaBee relied on the kernel version to decide which eBPF code
to compile and load. This method proved unreliable since many distros
backport features to old kernel versions. In order to make sure that
our code always builds correctly, this commit strips vmlinux.h from
the respoitory and adds code to the build.rs script to generate
vmlinux.h and detect features in it.

This enables our test cases to pass on rocky 9 linux.

Signed-off-by: Alan Wandke <adwandk@uwe.nsa.gov>